requirements.txt
scrapling>=0.4.7
python-dotenv>=1.0.0